+++ lib/altive_chat_room.dart
@@ -1,2 +1,3 @@
 export 'src/altive_chat_room.dart';
+export 'src/chat_message_card.dart';
 export 'src/models.dart';